requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pillow>=9.0.0
chromadb>=0.4.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0